+++ konata.core/Msf/Packets/Tlv/T148.cs
@@ -5,13 +5,13 @@ namespace Konata.Msf.Packets.Tlvs
     public class T148 : TlvBase
     {
         private readonly string _appName;
-        private readonly long _ssoVersion;
-        private readonly long _appId;
-        private readonly long _subAppId;
+        private readonly uint _ssoVersion;
+        private readonly uint _appId;
+        private readonly uint _subAppId;
         private readonly string _appVersion;
         private readonly string _appSignature;
 
-        public T148(string appName, long ssoVersion, long appId, long subAppId,
+        public T148(string appName, uint ssoVersion, uint appId, uint subAppId,
             string appVersion, string appSignature)
         {
             _appName = appName;
@@ -31,9 +31,9 @@ namespace Konata.Msf.Packets.Tlvs
         {
             StreamBuilder builder = new StreamBuilder();
             builder.PutString(_appName);
-            builder.PutUintBE((int)_ssoVersion);
-            builder.PutUintBE((int)_appId);
-            builder.PutUintBE((int)_subAppId);
+            builder.PutUintBE(_ssoVersion);
+            builder.PutUintBE(_appId);
+            builder.PutUintBE(_subAppId);
             builder.PutString(_appVersion);
             builder.PutString(_appSignature);
             return builder.GetBytes();
